auth/requirements.txt
Django>=5.2.8,<6.0
djangorestframework>=3.15.2,<4.0
djangorestframework-simplejwt>=5.3.1,<6.0
django-cors-headers>=4.4.0,<5.0
django-ratelimit>=4.1.0,<5.0
django-redis>=5.4.0,<6.0
python-dotenv>=1.0.1,<2.0
Pillow>=10.4.0,<11.0
dj-database-url>=2.2.0,<3.0